refactor: 이름에 빈칸 추가하는 로직을 출력 부분으로 이동
