+++ src/test/java/domain/user/UserTest.java
@@ -1,6 +1,5 @@
 package domain.user;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
@@ -8,7 +7,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 class UserTest {
@@ -78,26 +76,6 @@ class UserTest {
             String name = "ab c";
             assertThatThrownBy(() -> new User(name)).isInstanceOf(IllegalArgumentException.class)
                     .hasMessageContaining(User.NAME_FORMAT_ERROR_MESSAGE);
-        }
-    }
-
-    @Nested
-    @DisplayName("입력받은 문자열 변환 테스트")
-    class ConvertNameTest {
-        @DisplayName("변환된 문자열이 최대 길이 제한과 같은지 확인한다.")
-        @ParameterizedTest
-        @ValueSource(strings = {"i", "am", "fun", "dino", "mango"})
-        void shouldSuccessConvertNamesEqualMaxLength(String name) {
-            User user = new User(name);
-            assertThat(user.getName().length()).isEqualTo(NAME_LENGTH_LIMIT);
-        }
-
-        @DisplayName("입력받은 문자열이 지정된 형식에 맞게 변환되었는지 확인한다.")
-        @ParameterizedTest
-        @CsvSource(value = {"i:   i ", "am:  am ", "fun: fun ", "dino:dino ", "mango:mango"}, delimiter = ':', ignoreLeadingAndTrailingWhitespace = false)
-        void shouldSuccessConvertNames(String name, String expected) {
-            User user = new User(name);
-            assertThat(user.getName()).isEqualTo(expected);
         }
     }
 }
